+++ ruby/lib/swagger_client/models/response_info.rb
@@ -31,7 +31,7 @@ module SwaggerClient
     def self.swagger_types
       {
         :'copyrights' => :'Array<String>',
-        :'took' => :'Integer'
+        :'took' => :'Float'
       }
     end
